fixed some bugs
